+++ src/Farmer/Builders/Extensions.fs
@@ -2,13 +2,14 @@ namespace Farmer.Builders
 
 open Farmer
 open System
+open Farmer.Arm
 
 type ITaggable<'TConfig> =
     abstract member Add : 'TConfig -> list<string * string> -> 'TConfig
 type IDependable<'TConfig> =
     abstract member Add : 'TConfig -> ResourceId Set -> 'TConfig
 type IPrivateEndpoints<'TConfig> =
-    abstract member Add : 'TConfig -> (LinkedResource * String option) Set -> 'TConfig
+    abstract member Add : 'TConfig -> (SubnetReference * String option) Set -> 'TConfig
 
 [<AutoOpen>]
 module Extensions =