+++ RELEASE_NOTES.md
@@ -1,12 +1,20 @@
 Release Notes
 =============
+
 ## 1.6.32
 * CLI: Include `--overwrite true` option when executing `az storage blob upload-batch` with Azure CLI 2.34.0 and above.
+
+## 1.6.31
+* WebApps: Fix flakey deployments of web apps with multiple custom domains.
+* Deployments: Fix `ResourceId` generation when using a resource with a template.
+* AzureFirewall: Supports availability zones
+* WebApps/Functions: Add support for vnet integration
 
 ## 1.6.30
 * WebApps/Functions: Specify connection string types
 * WebApps/Functions: Allow adding IP restriction string with CIDR
 * Application Insights: Support for Workspace-enabled instances.
+* VMs: Priority and Spot Instance Settings
 
 ## 1.6.29
 * CLI: include `--only-show-error` option when executing Azure CLI commands.